Feat: 북마크 토글 추가, 좋아요 로직 추가

+++ src/domain/comment/components/CommentItem.tsx
@@ -4,7 +4,7 @@ import reportModalStore from "@/domain/report/stores/reportModalStore";
 import DropdownButton from "@/shared/components/DropdownButton";
 import tw from "@/shared/utils/tw";
 import Image from "next/image";
-import { useState } from "react";
+import React, { useState } from "react";
 import { useToggleReact } from "../api/useToggleReact";
 import { CommentType } from "../types";
 import { useQueryClient } from "@tanstack/react-query";
@@ -13,6 +13,7 @@ import TextBox from "@/shared/components/TextBox";
 import Button from "@/shared/components/Button";
 import { useUdtCommentById } from "../api/useUdtCommentById";
 import { useDelCommentById } from "../api/useDelCommentById";
+import Swal from "sweetalert2";
 
 interface CommentItemProps {
   item: CommentType;
@@ -46,10 +47,10 @@ function CommentItem({
     () => item.reactionCount
   );
   const [selectedOption, setSelectedOption] = useState<string | null>(null);
-  const setReportModalOpen = reportModalStore((state) => state.setIsOpen);
   const [isRepliesOpen, setIsRepliesOpen] = useState<boolean>(false);
   const [isEditMode, setIsEditMode] = useState<boolean>(false);
   const [editedContent, setEditedContent] = useState<string>(content);
+  const setReportModalOpen = reportModalStore((state) => state.setIsOpen);
 
   const { mutate: toggleReactMutation, isPending } = useToggleReact({
     onSuccess: () => {
@@ -70,6 +71,26 @@ function CommentItem({
   const { mutate: updateCommentMutation } = useUdtCommentById(feedId);
   const { mutate: deleteCommentMutation } = useDelCommentById(feedId);
 
+  const handleEditSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+
+    Swal.fire({
+      title: "수정",
+      text: "댓글을 수정하시겠습니까?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "수정",
+      cancelButtonText: "취소",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        updateCommentMutation({ id, content: editedContent });
+        setIsEditMode(false);
+      }
+    });
+  };
+
   // 이곳에서 드롭다운 메뉴 이벤트가 처리됩니다.
   const handleOptionClick = (option: string) => {
     setSelectedOption(option);
@@ -78,7 +99,20 @@ function CommentItem({
         setIsEditMode(true);
         break;
       case "삭제":
-        deleteCommentMutation({ id });
+        Swal.fire({
+          title: "삭제",
+          text: "댓글을 삭제하시겠습니까?",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "삭제",
+          cancelButtonText: "취소",
+        }).then((result) => {
+          if (result.isConfirmed) {
+            deleteCommentMutation({ id });
+          }
+        });
         break;
       case "신고":
         setReportModalOpen(true);
@@ -108,7 +142,10 @@ function CommentItem({
           {/* 유저이름, 컨텐츠 */}
           <div className="flex-1">
             {isEditMode ? (
-              <form className="flex gap-2">
+              <form
+                className="flex gap-2"
+                onSubmit={(e) => handleEditSubmit(e)}
+              >
                 <TextBox
                   className="flex-1"
                   initialValue={content}
@@ -172,7 +209,7 @@ function CommentItem({
         </div>
 
         {/* 대댓글이 있다면 재귀적으로 호출 
-          TODO: 답글에 답글까지는 불러오지 못 하는 것으로 보임. 개선 필요
+          TODO: 답글에 답글까지는(depth 2 이상의 replies) 불러오지 못 하는 것으로 보임. 개선 필요
         */}
         {replies && replies.length > 0 && isRepliesOpen && (
           <ul className="pl-2 mt-2 flex flex-col gap-2 border-l-2 border-slate-500">
